rename module to conventional format
