+++ src/calibration/bundle_adjustment/point_data.py
